Allow all position prompts to be aborted (from DynaHack). This applies to: jumping, polearm pounding, grappling hooks, scrolls of stinking cloud, paying with multiple shopkeepers in sight and spells of fireball and cone of cold.
All of these code paths had the ability to handle the position choice
being aborted by the player, but the `force` argument to `getpos()` was
preventing that from ever happening.

elunna: Note - this doesn't prevent the spell of jumping or directional spells from using up energy. Should this be addressed? MP is currently used up as soon as the spell is cast, but at least the turn won't be used with this commit.