Make the Castle drawbridge not always close with the passtune (from xNetHack). From UnNetHack, which got it from Spork. It's nice to have the castle meat grinder slightly less than completely reliable. (A plan I had to make a jammed drawbridge stay that way until fixed or destroyed got scrapped because there's currently nowhere good to store the data.)
Unlike in UnNetHack, this only happens when using the passtune; closing
magic always works. This is because I noticed when implementing it like
UnNetHack does, zapping closing at the drawbridge and having it fail to
close produces two "it gets stuck" messages and then erects a door on
the spot where the drawbridge ought to be. I don't want that behavior,
so just handwave it as closing magic preventing it from getting jammed.