Luck timeout is slowed instead of prevented by luckstones, based on their BUC status. This is originally from UnNetHack, but the timeout formula has been modified to timeout much faster than originally coded.
I thought the UnNetHack system of timeouts was a bit too generous. Under that system, it would take almost 100000 turns to timeout from +3 luck to 0 luck, and almost 200000 turns to go from +10 to 0.

I have revised it to be a bit more impactful, but still have fairly generous timeouts at the lower levels. To go from +1 luck to 0 is about 10000 turns, and I used this as the baseline.

- To go from +3 to 0 is about 35000 turns.
- To go from +6 to 0 is about 51000 turns.

If a player has a very low base luck (maybe from a future nerf or strange item), the timeouts can get quite short near the upper levels, but this is by design, otherwise the penalty is too easy to waive.