Buff the effects of the scroll of light (from xNetHack).
Uncursed light now gives radius 11 light (in line-of-sight, the same as
any other radius effect), and blessed light now illuminates the entire level.

The intention of this change is for the scroll of light not to be useless
like it currently is. With the current light rules, the scroll is
inferior to both the wand of light or the spell of light; hopefully
full-level, non-line-of-sight lighting will have good uses.