Raise odds of a cursed wand exploding (from xNetHack).
From UnNetHack. The intent here is to move this event from the "so rare
it's generally considered safe" tier to the "fairly rare but an actual
concern" probability tier in players' minds.

Un doesn't have the Cursed Wand Backfire Patch in effect, but xNetHack
does, so this slightly raises the overall odds of something bad
happening when a cursed wand is zapped -- but since the explosion check
is before the zapping-yourself check, the odds of the latter case are
actually smaller now.