Feedback when using unskilled or restricted weapons, or non-weapons (from EvilHack).
From both SporkHack and UnNetHack (modified) - if the player
tries to hit a monster with a weapon they are restricted in, or unskilled in,
or any non-weapon object, they'll never have better than a 75% chance of
landing a hit. The feedback generated when attacking under any of the
above-listed scenarios will vary (unskilled vs restricted vs non-weapon).
Bare-handed attack is exempt.