Revert the majority of level flips. I believe this is one of my (and many other people's) sticking points with 3.7 so far. There is just something about encountering familiar levels that you have seen for years and years, just flipped around and feeling different.
I've thought about this for a while, and I think it just has something to do with how it's comforting to have some anchor points in this RNG infested game. If EVERYTHING is RNG based, the player has nothing to hold onto. Psychologically, I believe this is pretty powerful and has been underestimated by the dev team. In the first iteration of Hack'EM, I also underestimated this factor in many places. This applies to all types of players: newbs, veterans, and especially speed runners. Just knowing where the entrances and exits of levels are located is very reassuring and worth preserving. I believe the introduction of flipped levels disrupts the normally calming process of proceeding through the dungeon, without much benefit to the player.

Therefore, my solution is to disable flipping for all levels except Sokoban and the Big Room. These two levels are designed for variety and lend themselves well to flipping, even if some people don't like the extra mental exercise - it's good for the brain!