Players cannot put known wands of cancellation or magical bags into bags of holding anymore (from dNetHack).
Once a player has identified these objects, it's pretty universally expected that they will *never* intentionally put them into their bag of holding. So, let's help them out with that.

I think this does some positive things for the game: 1) People won't rage quit and they will be happier. 2) Devs will see players not rage quitting and be happier. 3) Many promising games won't be wasted because of a simple mistake.

I believe there are more interesting challenges for the players to face than this, so it's one I'd like to see put to bed. Keep in mind, that if players are careless with their unidentified equipment, that is entirely on them.