Players are prevented from tipping known explosive items into known bags of holding.
They will simply flick them away before they enter the bag.

This change is meant to parallel the preventative behavior for putting
wands of cancellation and bags of tricks/holding into your bag of holding.
It's also intended to make the #tip command easier and less dangerous to use,
especially in the late game when players tend to have all of these items
identified.