Allow empty /oCancellation into BoH
Authored by hatcatter.

This is already allowed for bag of tricks/rats if they are uncharged.
The bag of tricks/rats implementation checks for amount of charges
and allows if there are none even if the player does not know this.
Since this information leak is allowed for bag of tricks/rats then this
should be fine for wand of cancellation too.

Reported by mobileuser.