Added diminished effects to several potions if they are diluted. I adapted this from EvilHack commit df086f65d1, which in turn got it from SporkHack.
Effects of potions are roughly cut in half if they are diluted. In EvilHack, some of the potions halve the upper part of rn1 calculation, but I updated it to just cut all the ending calculations in half. Potions of healing/extra healing/full healing grant half the max HP if diluted.