Rocks can be broken to produce flint stones (from xNetHack).
This combines commits b5a46e38b, 630b771be, 66f6a5313, 9570245ac, and 6b03fa9f5 from xnh.

Quoted thus from the first commit:

'The player can apply or rub rocks on other rocks or gray stones to bang
them together, breaking the rock and possibly creating some flint
stones. The number of flint stones produced is usually 0, but it helps
boost the chances of getting one or more if you are rubbing the rocks on a
touchstone or you are a Caveman.'

The rest of the commits are minor fixes.