Taming is more difficult on the Astral Plane.
This is something I first implemented in Hack'EM to counter the Astral Rain strategy. However, I set it a bit high (1 in 10 chance of taming), so here we are going to use a 1 in 5 rate so that players still have access to the strategy, but they will need to be prepared with extra resources to execute it.

Currently monsters cannot wear rings of slow digestion, so we can't implement the measures that EvilHack took (giving random astral denizens rings of SD to prevent purple worm engulfing). If that gets implemented, we might have to reexamine the probability of taming again.