Monsters can use magic flutes.
From xNetHack commit 9bb50d7ef8:

'Taken from SpliceHack, but with changes. Other monsters in the area will
also be put to sleep, not just the hero (the flutist is immune); the
sleep time for the hero is also the same as that of
put_monsters_to_sleep. Like SpliceHack, they will not play the flute if
you are already immobilized somehow.

Externify put_monsters_to_sleep and add a parameter for who is putting
the monsters to sleep (to center it on them and provide immunity).'

The xnh commit mentioned that monsters could not pick up magic flutes, but on testing I found that a large kobold easily picked one up and used it on me so that must have been updated recently in Vanilla.