Implement giving and taking pets' items via #loot
From xNetHack commit d9033e260:

'This is a feature I've wanted for a long time, though not really so much
for actual gameplay as for making various tests in wizard mode easier.
It's a lot simpler if you can just hand the goblin a wand of death
rather than setting it on the floor and watching him faff around the
room waiting for him to pick it up.

Nevertheless, it is an important gameplay feature! It greatly simplifies
the process of equipping a pet with the gear you want, because you can
now just take the gear you don't want away from them and give them the
appropriate gear. Formerly, the way to do this was drop items with
superior enchantments near them and hope their AI decided to swap it
out. A perennial problem was e.g. a pet Archon deciding a +2 club was a
better choice than their +0 Sunsword. (Note that this doesn't change the
AI; it'll still pick up and wield the club - but now the player can take
and keep it away from them.)

I wanted this to be done via #loot rather than copying the SpliceHack
implementation of #give, which adds a new command, only allows giving
one item at a time (and doesn't allow taking), and allows giving to
peacefuls, which opens up other edge cases for things like priests and
shopkeepers. I don't see a compelling need to allow giving/taking from
peacefuls right now, and plus, #give carries different connotations
(it's portrayed as a generous gift, which doesn't really work if it goes
both ways).

Currently shop handling is just through an addtobill() when the hero
takes an item from a monster. I've tested several scenarios, but it
seems to work correctly for the most part. The hero can take an item
from a pet standing just inside the shop door, but they are charged for
it, so if they try to walk off without paying they'll trigger the Kops.
If they take an item from a pet standing *on* the shop door from
outside, they're not charged. So this may ultimately make stealing
slightly easier, but hopefully not too much.'

I also included the follow-up commits to stay up to date with xNetHack:
4f0dbc370 - Fix: relobj use-after-free issue
9c9a8ff79 - Fix: m-prefix looting a container
82b3e9f55 - Fix: memory leak when looting monster
543ae7cb1 - Fix: charge to #loot pet outside shop
5b8e9f204 - Fix: peacefuls would never take any items given via #loot
141eba6bd - Handle welded equipment in #loot more consistently
7c3fbfbfb - Fix: giving pet wielded/quivered item via #loot
856de6378 - Delete saddle-specific #loot option
b5f2e9127 - Fix: pet unable to "return" item to shop
5a809772d - Fix: crash when #looting partial stack of weapons
f78758484 - Fix: removing saddle from pony via #loot/'take'
5672f38f6 - Permit giving an item to a saddled pony
4c9606d15 - Improve feedback for giving partial stack to pet
e0c10789d - Block giving monster your iron ball while punished