Farlook enhancements: show more basic information
From xNetHack commit e6ffadba5:

'Partially based on a feature in EvilHack to show when a monster is
equipped with armor and a weapon (but only a wielded weapon - if they
have a stack of darts that doesn't register).

Additionally, I implemented showing when a monster is fleeing. This may
end up giving the player a little too much information - prior to this
commit, you had to treat a monster you had previously sent fleeing as if
it might have recovered at any moment and is about to press the attack
again. But if the monster is fleeing, that *ought* to be one of the
things that's immediately obvious.'