Give flight to a number of major demons
From xNetHack commit cee37da93:

'The primary reason for this change is that a lot of demons have wings
and flight powers in the D&D versions they derive from, but strangely
don't in NetHack. The _impetus_ for this is that there's going to be
rather a lot of unsolid terrain in Gehennom (as opposed to the vanilla
maze version of Gehennom), and demons that can only move around on solid
floor have a disadvantage against a hero who surely will have levitation
or flying by then.

Affected demons are horned devils, erinyes, vrocks, bone devils,
nalfeshnees, and pit fiends. Balrogs and foocubi could already fly prior
to this.'