Add descriptive hit messages
From xNetHack commit 62a09cd14d:

'This is based on a patch originally by L, and then subsequently included
in several variants. This is based primarily on Fourk's version, though
I also incorporated bits of EvilHack's and SpliceHack's versions.

The reason for doing this is to spice up combat messages. In NetHack,
you see a LOT of "You hit foo! Foo hits!" over the course of the game,
and this makes it a bit more interesting. Note that if flags.verbose is
off, it'll all still be the dull messages, but I don't know if anyone
actually plays with it off, since that does things like suppress what
monster you're hitting and turn all uhitm combat into "You hit it".

Most of the new functionality is in two new functions; one for hits with
weapons and one for hits without. However, it also changes bite attacks
made with a beak into "peck", and adds an apparently missing case for
mhitm kick attacks (which exists in uhitm and mhitu).

One part of this that is unique to xNetHack is that it gathers a pool of
possible weapon hit verbs, and unevenly weights the chance of choosing
one. This happened because I liked the selection of verbs in Splice and
EvilHack, but didn't like some of them like "gore" or "lacerate"
happening too often.

Also, this uses vtense() to conjugate the verbs. Fourk is correct in
that it's icky and bad to use makeplural() for them, but resorting to
manual strings like Fourk does seems unnecessary when we have a verb
conjugator function.

Martial arts attacks and their hallucinatory counterparts are from
Splice, with some problematic ones removed.'