Fix: potential use-after-free when monster threw lit oil
From xNetHack commit 1e736ab724:

'Specifically at melee range, and when the monster was carrying more oil
that could ignite and explode. The code previously assumed that if there
was more than 1 oil in the stack to throw that it would definitely still
be around after the explosion to read the memory contents of, which was
not true.

Another ASan catch.'