Fix: vapors from thrown potion of monster detection could cause panic
From xNetHack commit 4111b7387:

'Another fuzzer-found crash. This is an indirect result of adding potion
vapor effects for every potion (vanilla monster detection potion vapors
do nothing). It turns out that strange_feeling() which can result from
some potion logic actually calls useup(), and this directly crashes when
the item in question is flying through the air, because it's not on any
object lists at the time.

There were several possible solutions, but the best seemed to be to put
the check for flying objects into strange_feeling() itself.'