Silver items can corrode.
From xNetHack commit 8c4a200aa:

'Cue the wailing and gnashing of teeth. Yes, silver doesn't corrode
easily, but it will certainly corrode when exposed to strong acids, and
potions of acid and acidic monsters qualify.

This is more of a gameplay-focused change, though, so that silver isn't
quite as much of a totally awesome ubermaterial as it is now.'