Creating Excalibur will fix any negative enchantment on it.
From xNetHack commit 4231e739; based on a similar commit in EvilHack.

One difference here is that if you are receiving Excalibur through crowning, your god will set its
enchantment to +1 at the minumum. Note that Excalibur that is randomly generated (if that's even possible)
or found in bones can still be negatively enchanted.