Zombie corpses may auto-revive.
From xNetHack commit 168647a0fd:

'This is actually not a SliceHack feature; if anything it's closer to a
GruntHack feature. I started with the SliceHack zombie apocalypse
commit, intending to take the less drastic design features of it more or
less verbatim, but this commit does not do that yet. It establishes part
of the framework for making it happen, and adds the unrelated feature in
which the corpses of any old zombies can revive themselves.

The frequency of zombie corpse revival is less than GruntHack, I think,
with a substantial chance that it will rot or go moldy rather than a
near-zero chance, but I haven't played GruntHack in a while so I'm not
sure.'

I also incorporated the is_zombie macro since that wasn't yet in Vanilla.