Add withering status effect from mummies.
From xNetHack commit 749ac4bb:

'This is another component of the Arc quest overhaul, as well as a
general feature that's been waiting for a while. The general motivation
is to make the M class more distinct instead of slightly-tougher-than-
zombie undead that have only boring physical attacks and are effectively
just popcorn after the player is strong enough to withstand those. The
motivation for the Arc quest is that a quest that predominantly features
boring monsters like that would itself be boring.

SpliceHack and EvilHack have already implemented features similar to
this. The implementation is based on SpliceHack's, though I've fixed a
number of issues I had with it. There are three main components: 1) the
new damage type AD_WTHR, 2) the new player-side intrinsic Withering, and
3) the new monster-side field mwither.

Unlike in Splice/Evil, mwither is implemented with a full byte, so that
it can track the amount of turns remaining like a proper timeout. Since
the severity of withering is strongly based on how much time it has left
to ravage one's HP, I wanted this to be more symmetric than the existing
system of "the monster may randomly recover from withering on 1/5 of
turns". Withering simply would be much weaker on monsters than it is on
the player, and multiple wither hits and the difference between a kobold
mummy's weak wither and a giant mummy's strong wither would be
meaningless.

I decided to make withering a pretty tough effect: few things cure it
(not even life-saving, which only prevents the actual cause of death -
which would be HP loss - and only fixes immediately "impending doom"
like nearly-timed-out illness). In Splice/Evil you can happily wipe it
away with a unicorn horn, which trivializes it. I may try out a feature
that allows curing withering with some method more available than prayer
but not as ubiquitous as a unicorn horn.

It's also a little nastier than in Splice/Evil, losing d2 points per
turn. (Some of the issues I noticed are that those implementations drain
HP per monster action, not turn - so they inexplicably wither faster or
slower. Here it happens per-turn as seems correct.) I was unable to get
an implementation for monster withering while the player is off-level
without a bunch of weird side effects, so I guess that just gets lumped
in as one of the "levels you're not on are frozen" effects.

I tried to implement things in such a way that extrinsic withering
remains a possibility in the future; the idea of some artifact that is
powerful but withers you while it's in use is attractive.'

HACKEM note: I imported all the mummy wither attacks as is - but did not raise their difficulty levels as xNetHack. I want to see how the game plays out with them as is and avoid adjusting difficulty where possible. If too high, we can back the wither attacks down to EvilHack levels (that sounds funny to say but the wither dice are lower in Evil).