Certain scared monsters can scream in fear.
From EvilHack commit 2ba242d39:

'From SporkHack, modified - there's a 20% chance that if a monster
becomes scared and wants to flee, it will scream with fright. Some
conditions - the monster has to have a voice, and can't be mindless or
silenced. If the scared monsters screams, there's a chance it could
wake up sleeping monsters nearby.'