Finite altar nerfs.
This originally comes from SpliceHack, but I made some adjustments while playing with it in Hack'EM. I wasn't satisfied with how it functioned in Hack'EM, so I took the opportunity to modify how this nerf operates. Altars are quite potent, and it's easy for a player to endlessly farm artifacts, gifts, luck, nutrition, and other resources from them. In my opinion, there should be some limits on the resources that can be obtained from an altar. Other dungeon resources like sinks and thrones have well-defined and short limits, and I believe this is a missing aspect of the altar feature.

The limitation is as follows: once the player has received two gifts or has been crowned, altars start to have a 50% chance to crack with each subsequent gift. There are two stages to the cracking now. The first stage is just a partial crack, which is purely cosmetic and doesn't affect the altar's functionality. However, receiving another gift from a cracked altar will certainly destroy it, losing it forever.

A few updates were carried over from Hack'EM. Altars on the Astral Plane will never be destroyed (thanks to NoisyToot for this). Note that they can still get cracked (for a little endgame drama) but are safe from destruction.

Being crowned also acts as a gift. So, if you are crowned on a cracked altar, it will also destroy it in the process.

I made being crowned act as having infinite gifts because crowning is a monumental act. The player gets a full set of resistances and usually an excellent weapon. If they aren't ready to gallantly venture forward after that, I think they need a kick in the butt!

There may be follow-up commits to possibly add another special level with a guaranteed altar.