Centaurs will keep their distance from the player naturally.
From xNetHack commit 8b93c1569:

'This is a port of EvilHack's M3_SKITTISH behavior, which comes from
SporkHack. I did not port the M3_ACCURATE parts at this time. Rather
than implementing it as a M3_ flag, I noticed that the only monsters in
xNetHack eligible to be skittish were bats (which are permastunned) and
centaurs. So instead of is_skittish I just made a keeps_distance macro
which currently just checks if the monster is a centaur.

Specifically this piggybacks off the same code that keeps fleeing
monsters and leprechauns away from the hero, forcing it to happen for
keeps_distance monsters. So for centaurs, this means that rather than
firing a couple bolts at you before getting into melee range and
bludgeoning you with the launcher, they will stay back and shoot their
bolts at you unless you back them up into melee range.

Known issue: if you don't attempt to melee the centaur yourself, and it
doesn't have any ranged attacks, it will just sort of dance around
nearby and not do anything on its own. There doesn't seem to be a
preexisting way of checking if a monster has any ranged attack option
available.'