Reduce AC of dragon scales to 1 and dragon scale mails to 3.
This is inspired by UnNetHack commit 98bdcfb8. In that commit, the AC for scale mail was reduced to 5, but UnNetHack also doesn't have secondary intrinsics. I think for the powers they confer, they should grant much less AC - otherwise the first DSM is a virtual wall of defense that carries the player through the rest of the game.

Because worn armor has it's weight reduced by 25% I also doubled the weight to a paltry 80. So worn DSN will only weigh 60.