Water Elementals can now engulf and suffocate.
From xNetHack commit ee808b73e:

'This is a change that is ultimately from GruntHack, by way of Keith
Simpson. Attacks of the combination AT_ENGL AD_WRAP are now parsed as
suffocation engulfing. On the hero, this begins a suffocation timer of
5; on monsters it just deals the normal damage for that attack, unless
the engulfed monster is breathless.

Additionally, when a water elemental engulfs the player, d3 inventory
items are wetted, and water elementals are given a passive rust attack.

This contains a fix to the GruntHack bug in which escaping a suffocating
engulfer by teleporting either you or it doesn't actually clear the
strangulation timer, causing the hero to die even after being safe.

Other weird edge cases that either didn't translate to 3.6 or were never
implemented in Grunt:
* If you put on an amulet of strangulation, it doesn't set Strangled to
  6 unconditionally (this could be used to buy more time when engulfed by
  a suffocating monster). It only sets it to 6 if higher than 6, or if
  it was 0 previously.
* Remove the Grunt case where an amphibious() monster feels quite
  comfortable inside a water elemental, because amphibious() actually
  means "amphibious or breathless", which means that a lot of monsters
  like vampires got the "quite comfortable" message which didn't make
  sense. Instead just give the general breathless case.
* Track the old value of uswallow in gulpmu, and use this to avoid
  re-printing "It's impossible to breathe in here!" when you get hit by
  another suffocation engulf attack while already engulfed.
* Add suffocation-from-monster cases to check_strangling, so that
  polyself into a breathless monster while being suffocated inside an
  engulfer will work to save you.'

Note: In the original commit, gelatinous cubes could also engulf - however, I have left this out. I think the gelatinous cube is probably fine as is, and with the paralysis creates some pretty helpless situations for the player. Adding this mechanic only for the water elemental feels more fitting - and it makes the plane of water kind of terrifying without magical breathing.