Ported gray fungus from EvilHack/SlashTHEM.
This is basically the same version as in EvilHack, but the SlashTHEM version allows them a little mobility when a speed of 1 - so that was added.

In Hack'EM I added the small touch attack from SlashTHEM, but it didn't really do much - so I've replaced it with a more interesting stick attack that the lichen has. Occasionally you'll get stuck to this little bugger, and what will you do then?