Ported thing from below from SpliceHack.
This was another popular genocide target in Hack'EM, so it's making the cut almost solely on that fact. These things are basically uber-krakens, capable of spellcasting as well as the drowning.

I noticed that despite being level 30, they had no innate MR. I corrected this and raised their MR to a hefty 95. Even the feared arch-lich (base level 25), only has 90 MR, so it makes sense to give it a bit more. They can still be genocided, but on the plane of water we don't want players to easily tame these beasts.

I replaced some monsters on the plane of water with them.