Added weretigers (from SLASH'EM).
There are a lot of were-foo in different variants. The weretiger feels like the best bang-for-our-buck that we can get. If we look at the werepanther in SLASH'EM, it is pretty similar to the werewolf. But the weretiger is a couple levels above, and can summon tigers which are also a big stronger than wolves.

The werebear from SpliceHack is a fun option, but requires adding a bunch of bear monsters. The werecockatrice and weredemon are also interesting, but then I feel we are bordering on too many werefoo. The same goes for the werespider and weresnake in SLASH'EM - so I think just the weretiger is enough.