Ported deepest ones from SLASH'EM.
I chose only to import the most powerful of the deeps. The deep ones and deeper ones are also ok, but don't have much of an effect. The deepest ones, however, are pretty nasty. One result of not importing the lesser deeps is that they can't be tamed and grown into the deepest ones.

Their stats are identical to SLASH'EM - but they have a few new tricks:
1) They are drain resistant.
2) They have flanking and will actively seek ways to flank you.
3) They cannot be genocided.