The iron ball and chain cannot be destroyed from  rusting.
This is also from EvilHack, where it is a measure to prevent the convict role from easily disposing of their starting ball and chain, but it's also good to make the scroll of punishment more potent. If we ever port in the convict it's a nice thing to have in place.