Giant spiders can ensnare monsters in webs (from EvilHack).
From EvilHack commit 5112589:

'Took some inspiration from dNetHack, created a new 'damage' type that on
a successful hit, entangles the target in a web trap. If the
target is already in a trap, the web is not created.'

I did not adjust the difficulty of the giant spider - players will have to just deal with that.