Don't let zombies dig out in Sokobon.
Inspired by hothraxxa's comments about the annoyingness of this feature.

To be fair, the rules of Sokobon usually are not disturbed in other places of the code (ie: giants are avoided when possible in the monster generation). I think this fits in with that and makes for a more enjoyable Soko experience.