Implemented MGC_FIRE_BOLT from EvilHack.
In EvilHack, this spell acts as a fireball. Unfortunately, this requires a lot of extra handling for things like blowing up shopkeeper doors and vaporizing fountains and requires a new type of damage type (MON_CASTBALL). I decided to go the easier route and just convert it to a single square effect. To be fair, this makes the spell quite similar to the pillar of fire spell, but just with much lower damage (depending on the monster level). Going this route also avoids situations where the caster also blows themselves up, so they can cast it at any range.

I also decoupled this spell from the ice bolt spell - which will be implemented in the next commit. In the future it might be worth making the spell depart further from pillar of fire, maybe by adding more physical damage. But for now, I think it's a decent way to empower the early game spellcasters.