Ported the acid blast monster spell from EvilHack.
Continuing with the trend from ice bolt and fire bolt, this no longer creates an explosion around the player - just a focused attack on the player's square.

This is one of the most feared spells in EvilHack, and because we don't really offer much for acid resistance, I lowered the blast to a d6 roll instead of d8. This is still very nasty with high level spellcasters, but I hope it will be a bit more manageable with this adjustment.