Ported stomping boots from SpliceHack.
This might be my favorite item from Splice... In Hack'EM I had further beefed it up so it could insta-kill small monsters but we are going to hold back on that and leave it to just always instakilling very small(tiny) monsters. With the addition of locusts, these might be worth wearing just to kill those.