Added the showdmg option from SLASH'EM.
This actually proved to be quite useful in Hack'EM for analyzing damage, looking at balance issues, and bugfixing. It can be enabled by setting the showdmg option to true/on.

For the time being, I will be leaving this feature in for players to use. If we settle in on a long term stable version, this might be turned off for non-wizmode use.

This initial commit covers most of the places that damage occurs, but in a future commit I want to consolidate the monster damage code to a damage_mon function and also make sure all of the other more obscure damage instances being done to the player (ie: while polymorphed) are also covered.