Ported the wand of poison gas from SpliceHack.
To make this wand a bit different, I have modified it slightly. This wand now shoots a ray similar to the wand of fireball from SLASH'EM - it will always end in an explosion of poison gas. This means that the ray also cannot be reflected nor will it bounce off of walls.

I originally didn't want to port this, but in order to port the wand of corrosion(acid), there needs to be a wand in between acid and lightning so there isn't a gap. Since poison clouds can rot items now, I think this wand might be quite a bit more dangerous especially combined with it's explosive (and vision impairing) qualities.

I also reduced the chance of items rotting in poison clouds slightly.