Added the wand of corrosion.
This wand was originally known as the wand of acid in SpliceHack. It appears in HACKEM with a few updates.

Notably, the ray from this wand (and now all acid streams) cannot be reflected. Since we are not importing the spellbook of acid blast, I think this makes this wand stand out more. It is very effective as most monsters do not resist acid, and it even penetrates reflection now. However, it will not bounce off walls as other rays (think of it as splashing into a wall). On the other hand, it is difficult to get a source of acid resistance, so this will also be a common chink in the player's endgame ascension kit. This makes yellow dragons both more dangerous and valuable for their scales. I do not plan on adding any more sources of acid resistance since that would work against the plan of making the endgame more difficult.