Fix: Zombies not reviving.
Some things got lost in translation while porting the reviving zombies from xNetHack. I was able to cross reference with EvilHack to get it right.

Because we also don't have 'Evil' zombies, I raised the chance of zombies reviving. Instead of the threat of illness/amnesia/poison, the player will just be threatened by big numbers - like a lot of zombie movies. Disposing of the zombies might be tricky since we don't have forges, but the tinning kit is always a good option.