Unicorn horn - offensive nerf.
As a d12/d12 two-handed weapon, the unihorn offers many weaker roles an easily attainable yet powerful weapon. Notably, archaeologists, cavemen, healers, and tourists. In Vanilla, priests can reach proficiency in unicorn horns, but in HACKEM, this was restricted to make the priest role more challenging. With the recent strength bonuses to two-handed weapons (NetHack commit 60c195685), this provides these roles with considerably more firepower. Even roles restricted in unicorn horns might consider them because of the +1 to-hit bonus, which slightly alleviates the -4 to-hit penalty of using an unskilled weapon. Although most roles will switch off the unicorn horn after acquiring a superior weapon or artifact, it remains useful for the rest of the game. It's also worth noting that HACKEM ported in the SLASH'EM mechanic where the effectiveness of unicorn horns depends on their enchantment - the result is that players will be enchanting their unicorn anyway, so getting a great weapon out of that is just icing on the cake.

The issue lies in the fact that the unicorn horn is already a potent healing item. For it to double as an effective offensive item makes most mediocre weapons like the quarterstaff, short sword, spears, sabers, and nearly all polearms obsolete. Even the respectable long sword (d8 vs. small, d12 vs. large) can't keep up with the unicorn damage-wise. The main drawback to the unicorn horn is that it requires both hands, preventing the player from wearing a shield. Since the roles that benefit most from the unicorn also can't two-weapon or might want to cast spells, it's not much of a drawback.

Strangely, the unicorn also outshines the crysknife (d10 vs. small, d10 vs. large), which doesn't make much sense. The crysknife is more challenging to obtain and requires more maintenance (and lacks any healing powers). They both also weigh the same.

I propose a solution to significantly reduce the offensive power of the unicorn horn. In EvilHack, the unicorn horn was changed to a one-handed weapon that deals d8/d10. However, I believe that still isn't enough since it outperforms many of the aforementioned weapons. I suggest modeling it after the great dagger from SLASH'EM, which is d6/d7 with +2 to-hit - this is roughly half of the unicorn horn's power. Nevertheless, I recommend keeping the +1 to-hit from the unicorn horn; otherwise, it encroaches on the territory of the crysknife. This change balances out the offensive capability of the unicorn horn, making room for other weapons to shine.