Wielded polearms grant +2AC.
This includes lances. This change was inspired by the way SpliceHack lets some weapons give AC bonuses. I didn't want to fully go that route because it requires adding a new acbon field to the weapons objects, but this is a minimal implementation which gets the job done. I had ideas for making the AC bonus scale with the weight or type of polearm, but decided to just keep it a simple, flat bonus similar to a large shield. A lot of people complain that polearms are totally useless, and they have some valid arguments in that. Hopefully this change gives a bit more power to the polearms.