Replaced vanilla sokoban levels with SLASH'EM levels.
I also added the Dragon of Bactria Sokoban ending from NetHack Fourk (that one has a nice surprise at the end).

This makes for a symmetrical array of levels:
* 3 variations for level 1
* 5 variations for level 2
* 5 variations for level 3
* 3 variations for level 4

With level flipping enabled (which it will be), this makes for a substantial increase in variety. I replaced the original Vanilla levels because I want people to experience a bit of SLASH'EM when they play HACKEM. I also really like those levels compared to the standard soko levels.

Also - I can't believe how painless it was to add these compared to the 3.6 des levels, long live lua!