Items that are 'lost' from the players inventory will no longer be un-identified.
This addresses a recent change in NetHack 3.7 where items that are stolen from the player or picked up after being thrown/dropped are considered 'lost', making the player lose their knowledge of the items enchantment, BUC state, and other details.

I believe this is a step backwards both for quality-of-life and for the mechanics of the game. As an example, I was playing a priest who had their starting mace stolen after 10000 turns of usage and play. After all that time, it seems rediculous that the player would just forget everything about a weapon they had been using for that long of a period. With the HACKEM mechanic of weapon familiarity, this effect isn't as annoying because one can relearn the status of a wielded weapon quickly. However, for projectiles this feature has been a pain from the beginning.

Recent efforts have also tried to implement a message "You learn more about your items by comparing them.", however this message is confusing both for new and old players and doesn't really fix the annoyingness of the feature.

I think that as a rule of the game, once you have identified something - it should stay identified. There was a lot of time spent on the problem of mind flayers and amnesia, and this felt like a step backward to the old amnesia mechanic.

I believe I fixed the main part - however, I am not 100% sure. I left the framework of the lost items in but just removed the merging effects. This seems to bring the item behavior back to normal for stolen items.