Endgame genocide nerf.
One popular tactic players use is to get to the plane of water then genocide dangerous drowning monsters like giant eels, electric eels, or mermaids. If a player genocides those monsters *before* reaching the plane of water, they are replaced by other monsters, but not when the genocide is performed on-site.

This change altars the behavior of genocide in all planes levels so blessed/global genocides are no longer possible. A regular/uncursed genocide takes the place of a blessed genocide and is only 25% as effective. Players will still be able to do blessed genocide before reaching the planes as usual. Cursed genocides are not affected.

This might require further testing, and the message shown when level-wide genociding might need tweaking. My hope is that when players do decide to genocide lots of ; monsters, they will be replaced by deepest ones which are quite nasty in swarms and also ungenocideable.