Imported the M3_ACCURATE tag from EvilHack.
I think this is very much needed because once your character gets to lowish AC, a lot of projectiles just miss. In particular, elves seem to miss a lot when they really should not, so I think Evil really hit it where it needed it here.

One class of monsters I also buffed with accuracy was piercers.