Troll meat provides temporary intrinsic regeneration.
From xNetHack commit 23c6e3372:

'Note by copperwater: I reimplemented much of this, because I'd prefer
not to enumerate troll species in case any more ever get added. Better
to do it based on the monster class.

In addition, I reduced the duration of the regeneration you get to 2d6.
Displacer beasts are fairly rare and 6d6 displacement is fine from them,
but trolls are common and also are commonly tinned to stop them from
reviving, so it's not hard to get a bunch of blessed troll tins.'