Buff war hammers into a competitive two-handed weapon.
From xNetHack commit 0b04b8fed6:

'This ups their damage to 2d6 small / 2d8 large. The primary reason for
this change is because d4+1/d4 is ridiculously low and not indicative of
the smashing power of a hammer.

Affected artifacts are Mjollnir and Ogresmasher, with the main intended
beneficiary being Ogresmasher. People have pointed out some nitpicks
with this change's effect on Mjollnir previously:
1) It makes valkyries harder since they can no longer wield it in one
   hand and wear their starting shield in the other. This is fine by me;
   it makes for a more interesting choice whether to forego the shield
   or not, especially since Mjollnir is now better than Excalibur
   damage-wise to non-shock-resistant targets. (Possibly it's too
   powerful, and the 1d24 damage needs to be tuned down a bit.)
2) Norse mythology is very specific about it being short-handled and
   wielded by Thor in one hand. This can be handwaved away as its being
   the hammer of a god, and not necessarily existing at a mortal scale;
   alternatively, due to its famous weight, it's all a mortal can do to
   wield the one-handed weapon with both hands.'

ESL: I think this also complements the changes to unicorn horns and the priest quite nicely. Priests always get Mjollnir for a crowning gift, and this gives them something with a bit more oomph. Valkyries won't have any problems as they just always go for long sword artifacts anyway. This also can serve to take the place of unicorns as a better two-handed weapon, however it may not be easy to find one, it weighs more, it's not erodeproof, and your role may or may not be able to attain proficiency with it - more balanced though!