Remove monster maximum difficulty limit with demigod flag set.
From xNetHack commit afbf6a547b:

'From UnNetHack; the reason being that in the ascension run things like
being on a low dungeon level or getting drained to a low experience
level shouldn't have any effect on the toughness of monsters that
generate.

Unlike UnNetHack's implementation, this uses a constant 256 as the new
max difficulty (monster difficulty is stored as a uchar) instead of
using Demogorgon's difficulty, because it's plausible that someone goes
and makes an even scarier monster at some point. (Not that I have any
plans to.)'