Only elves can safely enchant elven armor over +3.
Two reasons I'm restricting this: 1) Reduce the endgame AC for most of the cast. 2) Make elves have a nice distinguishing characteristic.

We could also use a lore based excuse that only elves are familiar or careful enough with their equipment to be able to over-enchant it - yes, that will do.

To help the unknowning non-elven players, I made the armor always vibrate when it's above the safe point.