Primary spellcasters get a memory boost when they cast spells.
The roles that actually start with spellcooks (healers, monks, priests, and wizards) get a 500 turn boost to their successfully cast spells.

This change was originally in SLASH'EM, but I folded in some small things from the way EvilHack does it. There is a requirement for at least 6 INT, otherwise no benefit is given.

The maximum turn count was 70000 turns in SLASH'EM but I think twice the KEEN amount is fine. In vanilla this is 20000 normally, but I plan on changing this to 10000, making the maximum remembrance 20000 turns.