Spellcasting KEEN is 10000.
This is meant to be a soft nerf to spellcasting for non-spellcasters. In the previous commit, we gave primary spellcasters a bonus to memory for casting spells. This is the followup to make that actually matter.

For non-spellcasters, they will have to be much more vigiliant in maintaining their spells and spellbooks - or just focus on their strengths instead.