Implement peaceful monster underlining in tty and curses.
From xNetHack commit 29b2c92d5a:

'Create a whole new set of glyphs for peaceful, non-pet monsters, and add
tty/curses code to pick up on them and turn on the underline attribute
for those monsters.

Add a new option "underline_peacefuls" and iflags value to enable the
player to turn it on or off.

I *think* I've done the peaceful worm tails correctly? Really can't tell
with worm code.

During testing, discovered that the wizgenesis command didn't render
the underline until the monster moved; this was because it set the
monster to peaceful after all the newsyms were called on that monster.
More instances like this might exist; they will probably have to be
handled on a case-by-case basis.'