Make training weapons out of unskilled easier.
The UnNetHack skills system raises the required hits to 100, but this feels quite restrictive and can be dispiriting to train up, especially when unskilled weapons hit less and projectiles randomly go off course. We will cut the requirement in half so it's only 50, This is still more than double what it is in Vanilla but should at least open the door for many more opportunities. The real training begins at Basic skill.