Tone down berserkers a bit.
Instead of a 50% chance each turn of going berserk (below 1/2 HP), they have a 20% chance. I want this to be a frequent thing, but not an all-the-time thing. The player should usually get a couple of turns to kill monsters before they berserk, but occasionally be surprised and have to back off.