Ported forges and lava demons from EvilHack.
Currently the implementation mirrors that of EvilHack except for the #forge command. I am leaving that out because I want the simplest version of this feature as possible. Forges let the player repair metallic armor, bless items, dispose of zombies or organic items, dispose of the iron ball and chain, and summon lava demons and fire elementals. The actual forging mechanics complicate matters more than I want and introduce a new command. I'm also not sure how to balance out the forging recipes, what to allow, what quantities make sense, and etc etc etc - so I'm just going to bypass it. The core features of the forge are fantastic and I think we can actually add a couple more subtle ones to really make them shine.