Revert 218c0d4a3 - Stinking clouds block line of sight
This is just a soft revert, I only disabled the code that blocks the vision but kept it in for future compatibility. Being inside a poison cloud still blocks vision though.

My reason: I personally don't like this change. Not seeing the poison clouds makes strategically moving around them more difficult - I know this is the intended purpose, but we can also see all the boulders in Sokoban because it's just a video game. Seeing the effect of your stinking cloud or wand of poison gas feels good and is More Fun. Not seeing it makes it feel like it didn't happen. My reasons are feeling based, but then again, so are a lot of video games.