Fix: hatching wielded/quivered egg called weight()
From xNetHack commit a6e359aeb:

'According to vanilla code, it is not a bug for setnotworn to be called
with a hatching egg that has already had its quan changed to 0. So this
fixes it inside setworn, only recalculating the object's weight if it
was worn armor. (If someone manages to wear an egg as armor, that will
be a worse bug than this.)'