Branchport always brings you to the entry level of a branch.
Reverting the 3.6.1 change that made branchport put the player in the deepest level in the branch. This is mostly a quality-of-life change because a key use for the EotA for me is to quickly go back to my stash at Sokoban. If I have to then navigate down 4 levels to get to my stash it makes it much more annoying. I'm curious as to the root cause of this change, but I fundamentally disagree with it. I think it makes the behavior less consistent.