Tone down defensive monster behavior.
Previously, if a monster was surrounded, it would try to use a defensive manuever to get out of the situation. However, this results in a lot of unnecessary actions like monsters reading scrolls of teleport. There was no HP or other checks so the monster was purely looking for a new place to move.

I added a check so that the monster has to either be fleeing or hit a 1 in 13 chance when surrounded to do something evasive, hopefully most of the time they are able to find a new spot so they won't waste so many items.