Define all themed rooms with unusual floor features as type themed.
From xNetHack commit f0b18cae:

'Doing this makes them ineligible to become special rooms. This is to
prevent a bunch of potential shenanigans where some assumption is made
in the special room generation logic that the room will be regular and
ordinary, but it's not, and then things break.

There's a github issue for vanilla about this (with the picture of the
shop that spawned in the 3x3 lava room), but I don't think the devteam
is going to fix it upstream.

Note that it doesn't only require unusual floor terrain - I excluded the
devnull pool challenge room because it has traps, and the buried
treasure room because it puts items under the floor that a clever hero
is supposed to interact with.'