Reduce special themed rooms under difficulty level 5.
This was a key feature of Hack'EM and I hope to continue it on in HACKEM. I want the first 5 or so levels of the game to feel like traditional NetHack, with no new themed rooms generating. This is to gradually introduce people to the new variant with a feeling of nostalgia or comfort of the old school NetHack level generation. As the player progresses they should experience new and exotic rooms.

As implemented, it looks like the LUA calculation uses level difficulty, which I believe factors in the player's level as well, however, I'm not sure. I do know that it does accomplish what I'm looking for so we will revise this as time goes on.