Piercers and lurkers/trappers should always generate hidden.
I noticed that sometimes going downstairs to a new level that piercers would be visible. This change should make them always hide on creation, making them act more deceptive as intended.

To test this I had to modify the monster wishing a bit so that monsters can be hidden when wiz-mode created. This seems to have a strange flickering effect when they are summoned, but they do end up hidden so I will keep it that way unless it has some effect on gameplay.