Redcaps can get different gear, berserk, and leave no corpses.
This is just based on a cursory glance at some D&D lore, but these updates should make the redcaps even more tough than they already are. I also wanted to make sure they got their traditional scythe that was granted in SpliceHack, but also allow them to start with a sling for ranged attacks.

They are known to be bloodthirsty savages, so berserk seems necessary - but they will be very dangerous now, we'll see how players deal with them...