Grease traps.
Based on the ideas from https://nethack-yanis.github.io/yanis/4083.html, authored by aosdict.

This is a fun (but probably annoying) trap that pushes the boundaries of what grease can do. I modeled the effects off of the rust trap, so there is a kind of symmetry between the two traps, but they have vastly different functions.

Players who walk into grease traps can have a blast of grease hit them. This has some different effects:
1) Sprayed on the head. If you are wearing a helmet, it may be greased. You also get blinded by goop unless you are wearing a visored helmet. If you are wearing any blindfold accessories, they will slip off and possibly get greased. When you are blinded this way, it counts the same as being blinded by a cream pie and can be wiped off by a towel.

2) Sprayed on the arm. If wielding a weapon or wearing a shield, it can be greased and slip from your grip. You get the glib status for 10-15 turns.

3) Sprayed on the feet. If wearing boots, they can get greased and slip off (unless buckled boots).

4) Sprayed on your torso. Towels and quivered ammo will usually become greased. A random item can also become greased and slip from your pack. Your outermost armor will also become greased (but not fall off). If you are riding a steed, the saddle will become greased and you'll fall off.

In all cases, you get the fumbling status because we assume you stepped in a very greasey pool (as a result of much grease trap history). This status lasts for a short while, but can be cured by applying a towel to wipe off your feet.

Once it has been set off once and seen, the trap has a 1 in 15 chance of disappearing each use. It can also be #untrapped with a rubber hose resulting from a successful untrap.

Currently the handling for monsters stepping onto traps is more simple than for players. A random piece of armor will be greased but no other effects are implemented. This will be a good followup commit.