Unicorn skill and unicorn horn use.
Having skill in unicorn horn has a positive impact on your success. Basic skill grants a 10% bonus to success, skilled 20%s, and expert 40%. This allows roles that can attain proficiency to not have to enchant it so high.

Furthermore, successfully applying a unihorn also exercises your skill in it by 5 points. This lets a player rapidly advance from unskilled to basic skill (requiring only 20 uses).

I also gave the healer a little bonus. Instead of the standard 30% success rate when the unihorn is +0, they have a 40% success rate.